+++ components/31-molecules/contact-block/contact-block.config.js
@@ -10,17 +10,17 @@ module.exports = {
   collator: function (markup, item) {
     return `<!-- Start: @${item.handle} -->\n<dt>${item.name}</dt><dd>${markup}</dd>\n<!-- End: @${item.handle} -->\n`;
   },
-  context: {
-    title: 'Contact us',
-    description: 'Nulla nec eros lorem. Morbi at augue eu purus congue auctor. Sed id volutpat odio. Nunc congue diam vel diam venenatis, ut consequat urna dapibus. Vivamus at lobortis odio. Sed felis mauris, mattis eget purus at, pharetra ullamcorper risus.',
-    tel: '09 123 45 67',
-    email: null,
-    link: '#'
-  },
   variants: [
     {
       name: 'contact-block-with-phone',
-      preview: '@preview'
+      preview: '@preview',
+      context: {
+        title: 'Contact us',
+        description: 'Nulla nec eros lorem. Morbi at augue eu purus congue auctor. Sed id volutpat odio. Nunc congue diam vel diam venenatis, ut consequat urna dapibus. Vivamus at lobortis odio. Sed felis mauris, mattis eget purus at, pharetra ullamcorper risus.',
+        tel: '09 123 45 67',
+        email: null,
+        link: '#'
+      }
     },
     {
       name: 'contact-block-with-email',
@@ -31,6 +31,26 @@ module.exports = {
         tel: null,
         email: 'info@stad.gent',
         link: '#'
+      }
+    },
+    {
+      name: 'contact-block-without-description',
+      preview: '@preview',
+      context: {
+        title: 'Contact us',
+        tel: '09 123 45 67',
+        email: null,
+        link: '#'
+      }
+    },
+    {
+      name: 'contact-block-without-link',
+      preview: '@preview',
+      context: {
+        title: 'Contact us',
+        description: 'Nulla nec eros lorem. Morbi at augue eu purus congue auctor. Sed id volutpat odio. Nunc congue diam vel diam venenatis, ut consequat urna dapibus. Vivamus at lobortis odio. Sed felis mauris, mattis eget purus at, pharetra ullamcorper risus.',
+        tel: null,
+        email: 'info@stad.gent'
       }
     }
   ]