+++ components/31-molecules/opening-hours/week-month-widget/week-month-widget.config.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = {
+  name: 'Week Month Widget',
+  status: 'alpha',
+  preview: '@preview'
+};