+++ components/41-organisms/opening-hours-accordion/opening-hours-accordion.config.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = {
+  title: 'Opening hours accordion',
+  status: 'alpha',
+  preview: '@preview'
+};